BackwardColNum and types

+++ oneflow/core/kernel/add_kernel.cpp
@@ -9,7 +9,7 @@ void AddKernel<device_type, T>::ForwardDataContent(
   Blob* out_blob = BnInOp2Blob("out");
   const Blob* in_blob0 = BnInOp2Blob(this->kernel_conf().input_bns()[0]);
   out_blob->CopyDataContentFrom(ctx.device_ctx, in_blob0);
-  const int count = out_blob->shape().elem_cnt();
+  const int64_t count = out_blob->shape().elem_cnt();
   for (size_t i = 1; i < this->kernel_conf().input_bns().size(); ++i) {
     const Blob* in_blob = BnInOp2Blob(this->kernel_conf().input_bns()[i]);
     KernelUtil<device_type, T>::Axpy(ctx.device_ctx, count, static_cast<T>(1),
